Original files from 3.5" diskette accompanying a book of Bourne Shell examples.
The full name of the book is "Portable Shell Programming: An Extensive
Collection of Bourne Shell Examples" (ISBN 0-13-451494-7) by Bruce Blinn, and
published by Prentice Hall.

The format of the disk was unusual in that it didn't contain a normal file-
system, under which files were stored. Instead, the raw data bytes of a tar
archive containing the files were written to the disk. To get the files, you had
to enter something like:

	tar -xvf /dev/sdb